refactor: consolidate navigation imports and remove unused files
