+++ web/src/platform/routes/Admin/Database.tsx
@@ -12,8 +12,8 @@ import { fetchApiVoid } from '@/lib/api';
 import { useToast } from '@/hooks/use-toast';
 import { useApiQuery } from '@/hooks/use-api';
 import { useDeleteDialog } from '@/lib/utils';
-import { usePaginate } from '@/hooks/pagination';
 import { PostgreSQL } from '@/svg/PostgreSQL';
+import { usePaginate } from '@/hooks/pagination';
 import { platformApiPath } from '@/lib/platform-api';
 import { databasesQueryKey } from '@/lib/query-keys';
 import CreateDatabase from '@/components/dialogs/CreateDatabase';
