+++ web/src/platform/routes/Admin/Layout.tsx
@@ -1,8 +1,8 @@
 import { Outlet } from 'react-router';
 import { Auth } from '@/components/Auth';
 import PlatformLayout from '@/platform/layout';
+import { ADMIN_NAVIGATION } from '@/platform/navigation';
 import { PageContainer } from '@/components/PageContainer';
-import { ADMIN_NAVIGATION } from '@/platform/admin/navigation';
 
 /** Renders the authorized admin shell with tabbed navigation. */
 export default function AdminLayoutRoute() {
