+++ web/src/platform/routes/Admin/Storage.tsx
@@ -13,8 +13,8 @@ import { fetchApiVoid } from '@/lib/api';
 import { useToast } from '@/hooks/use-toast';
 import { useApiQuery } from '@/hooks/use-api';
 import { useDeleteDialog } from '@/lib/utils';
-import { storagesQueryKey } from '@/lib/query-keys';
 import { usePaginate } from '@/hooks/pagination';
+import { storagesQueryKey } from '@/lib/query-keys';
 import { platformApiPath } from '@/lib/platform-api';
 import CreateStorage from '@/components/dialogs/CreateStorage';
 import { DeleteConfirmation } from '@/components/dialogs/DeleteConfirmation';
