+++ web/src/platform/routes/Legal/Terms.tsx
@@ -5,7 +5,7 @@ import { Stack } from '@astryxdesign/core/Stack';
 import { Heading } from '@astryxdesign/core/Heading';
 import { List, ListItem } from '@astryxdesign/core/List';
 import type { ArticlePage } from '@/lib/articles';
-import { LEGAL_GROUPS } from '@/platform/legal/navigation';
+import { LEGAL_GROUPS } from '@/platform/navigation';
 import { ArticlePageRenderer, createArticleMeta } from '@/platform/routes/Articles';
 
 export const metadata = {
